extract urlHelper
